+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -3303,47 +3303,3 @@ def test_set_value(dict, paths, value, expected_dict):
     """
     set_value(dict, paths, value)
     assert expected_dict == dict
-
-
-from demisto_sdk.commands.common.tools import collect_all_inputs_in_use, collect_all_inputs_from_inputs_section
-
-
-from demisto_sdk.commands.validate.tests.test_tools import create_playbook_object
-
-
-playbook = create_playbook_object()
-def test_collect_all_inputs_in_use_empty(mocker, playbook):
-    """
-    Given:
-        - A playbook with no input usages
-    When:
-        - Running collect_all_inputs_in_use
-    Then:
-        - An empty set should be returned
-    """
-    mocker.patch.object(playbook, 'content_item_as_text', return_value='')
-    assert collect_all_inputs_in_use(playbook) == set()
-    
-def test_collect_all_inputs_in_use_with_usage(mocker, playbook):
-    """
-    Given:
-        - A playbook using input1 and input2
-    When:
-        - Running collect_all_inputs_in_use 
-    Then:
-        - A set with input1 and input2 should be returned
-    """
-    #mocker.patch.object(playbook, 'content_item_as_text', return_value='text {{input1}} more text {{input2}}')
-    assert collect_all_inputs_in_use(playbook) == {'input1', 'input2'}
-
-
-def test_collect_all_inputs_from_inputs_section(playbook):
-    """
-    Given:
-        - A playbook with input1 and input2 defined
-    When:
-        - Running collect_all_inputs_from_inputs_section
-    Then: 
-        - A set with input1 and input2 should be returned
-    """
-    assert collect_all_inputs_from_inputs_section(playbook) == {'input1', 'input2'}
